+++ apps/dashboard/src/app/(main)/(protected)/projects/[projectId]/sidebar-layout.tsx
@@ -227,7 +227,8 @@ function TeamMemberBreadcrumbItem(props: { teamId: string }) {
 function NavItem({ item, href, onClick }: { item: Item, href: string, onClick?: () => void}) {
   const pathname = usePathname();
   const selected = useMemo(() => {
-    return item.regex.test(pathname);
+    let pathnameWithoutTrailingSlash = pathname.endsWith("/") ? pathname.slice(0, -1) : pathname;
+    return item.regex.test(pathnameWithoutTrailingSlash);
   }, [item.regex, pathname]);
 
   return (
@@ -235,8 +236,8 @@ function NavItem({ item, href, onClick }: { item: Item, href: string, onClick?: 
       href={href}
       className={cn(
         buttonVariants({ variant: 'ghost', size: "sm" }),
-        selected && "bg-muted",
         "flex-grow justify-start text-md text-zinc-800 dark:text-zinc-300 px-2",
+        selected && "bg-muted",
       )}
       onClick={onClick}
       prefetch={true}
@@ -382,11 +383,21 @@ export default function SidebarLayout(props: { projectId: string, children?: Rea
 
   return (
     <div className="w-full flex">
-      <div className="flex-col border-r min-w-[240px] h-screen sticky top-0 hidden md:flex">
+      <div style={{
+        width: "160vw",
+        height: "180vmin",
+        position: "fixed",
+        top: "100%",
+        left: "50%",
+        transform: "translate(-50%, -10%)",
+        zIndex: -10000,
+        backgroundImage: "radial-gradient(50% 50% at 50% 50%,#4b5df7 0,#4b5df773 23%,#4b5df740 50%,#4b5df700)",
+      }} />
+      <div className="flex-col border-r min-w-[240px] h-screen sticky top-0 hidden md:flex backdrop-blur-md bg-white/20 dark:bg-black/20 z-[10]">
         <SidebarContent projectId={props.projectId} />
       </div>
       <div className="flex flex-col flex-grow w-0">
-        <div className="h-14 border-b flex items-center justify-between sticky top-0 bg-white dark:bg-black z-10 px-4 md:px-6">
+        <div className="h-14 border-b flex items-center justify-between sticky top-0 backdrop-blur-md bg-white/20 dark:bg-black/20 z-10 px-4 md:px-6">
           <div className="hidden md:flex">
             <HeaderBreadcrumb projectId={props.projectId} />
           </div>