+++ packages/stack-shared/src/utils/strings.tsx
@@ -119,6 +119,7 @@ export function deindent(strings: string | readonly string[], ...values: any[]):
 }
 
 export function extractScopes(scope: string, removeDuplicates=true): string[] {
+  // TODO what is this for? can we move this into the OAuth code in the backend?
   const trimmedString = scope.trim();
   const scopesArray = trimmedString.split(/\s+/);
   const filtered = scopesArray.filter(scope => scope.length > 0);
@@ -126,10 +127,14 @@ export function extractScopes(scope: string, removeDuplicates=true): string[] {
 }
 
 export function mergeScopeStrings(...scopes: string[]): string {
+  // TODO what is this for? can we move this into the OAuth code in the backend?
   const allScope = scopes.map((s) => extractScopes(s)).flat().join(" ");
   return extractScopes(allScope).join(" ");
 }
 
+export function escapeTemplateLiteral(s: string): string {
+  return s.replaceAll("`", "\\`").replaceAll("\\", "\\\\").replaceAll("$", "\\$");
+}
 
 /**
  * Some classes have different constructor names in different environments (eg. `Headers` is sometimes called `_Headers`,
@@ -218,8 +223,23 @@ export function nicify(
   };
 
   switch (typeof value) {
-    case "string": case "boolean": case "number": {
+    case "boolean": case "number": {
       return JSON.stringify(value);
+    }
+    case "string": {
+      const isDeindentable = (v: string) => deindent(v) === v && v.includes("\n");
+      const wrapInDeindent = (v: string) => deindent`
+        deindent\`
+        ${currentIndent + lineIndent}${escapeTemplateLiteral(value).replaceAll("\n", nl + lineIndent)}
+        ${currentIndent}\`
+      `;
+      if (isDeindentable(value)) {
+        return wrapInDeindent(value);
+      } else if (value.endsWith("\n") && isDeindentable(value.slice(0, -1))) {
+        return wrapInDeindent(value.slice(0, -1)) + ' + "\\n"';
+      } else {
+        return JSON.stringify(value);
+      }
     }
     case "undefined": {
       return "undefined";
@@ -251,7 +271,7 @@ export function nicify(
         }
       }
       if (value instanceof URL) {
-        return `URL(${JSON.stringify(value.toString())})`;
+        return `URL(${nicify(value.toString())})`;
       }
       if (ArrayBuffer.isView(value)) {
         return `${value.constructor.name}([${value.toString()}])`;