+++ docs/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @stackframe/docs
 
+## 2.7.12
+
+### Patch Changes
+
+- Various changes
+
+## 2.7.11
+
+## 2.7.10
+
+### Patch Changes
+
+- Various changes
+
+## 2.7.9
+
 ## 2.7.8
 
 ### Patch Changes